backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.3.1
Pillow==10.3.0
python-decouple==3.8
djangorestframework-simplejwt==5.3.1